+++ app/utils/MiscUtils.ts
@@ -1,24 +1,20 @@
-import { format } from "date-fns";
 import {
-  DIRECTION_SEGMENTS,
   FAHRENHEIT_BASE,
   FAHRENHEIT_TO_CELSIUS,
   HALF_PHASE,
-  KPH_CONVERSION_FACTOR,
   MAX_VISIBILITY_MILES,
-  MILES_CONVERSION_FACTOR,
-  MILLISECONDS_PER_SECOND,
-  MM_TO_INCHES,
   PERCENT_MULTIPLIER,
   POPULATION_THRESHOLD,
   QUARTER_PHASE,
   RAIN_PRECISION,
   SLICE_END_INDEX,
-  THREE_QUARTER_PHASE,
-  WIND_DIRECTION_DIVISOR
+  THREE_QUARTER_PHASE
 } from "../config";
 import { CurrentWeatherData } from "../models/componentProps";
-import { Alert, WeatherItem } from "../models/weatherTypes";
+import { WeatherItem } from "../models/weatherTypes";
+import { getMiles, mmToInches } from "./MeasurementUtils";
+import { parseTime } from "./TimeUtils";
+import { getWindDirection, getWindGust, toKph } from "./windUtils";
 
 /**
  * Returns the moon phase based on the given phase value.
@@ -64,84 +60,6 @@ const toCelsius = (num: number): number =>
   Math.floor((num - FAHRENHEIT_BASE) * FAHRENHEIT_TO_CELSIUS);
 
 /**
- * Returns the wind direction based on the given angle.
- * @param {number} direction - The wind direction angle in degrees.
- * @returns {string} - The wind direction as a string.
- * @example
- * getWindDirection(0); // "N"
- * getWindDirection(45); // "NE"
- * getWindDirection(90); // "E"
- * getWindDirection(135); // "SE"
- * getWindDirection(180); // "S"
- * getWindDirection(225); // "SW"
- * getWindDirection(270); // "W"
- * getWindDirection(315); // "NW"
- * getWindDirection(360); // "N"
- */
-const getWindDirection = (direction: number): string =>
-  [
-    "N",
-    "N/NE",
-    "N/E",
-    "E/NE",
-    "E",
-    "E/SE",
-    "SE",
-    "S/SE",
-    "S",
-    "S/SW",
-    "SW",
-    "W/SW",
-    "W",
-    "W/NW",
-    "NW",
-    "N/NW"
-  ][Math.round(direction / WIND_DIRECTION_DIVISOR) % DIRECTION_SEGMENTS];
-
-/**
- * Converts speed from miles per hour to kilometers per hour.
- * @param {number} speed - The speed in miles per hour.
- * @returns {number} The speed in kilometers per hour.
- * @example
- * toKph(60); // 96
- * toKph(30); // 48
- * toKph(45); // 72
- * toKph(100); // 160
- */
-const toKph = (speed: number): number =>
-  Math.floor(speed * KPH_CONVERSION_FACTOR);
-
-/**
- * Parses the given time value and returns a formatted date string based on the provided locale and timezone.
- * @param {number} time - The time value to parse (in seconds).
- * @param {string} locale - The locale to use for formatting the date string.
- * @param {string} timezone - The timezone to use for formatting the date string.
- * @returns {string} The formatted date string.
- * @example
- * parseTime(1620000000, "en-US", "America/New_York"); // "5/3/2021, 12:00:00 AM"
- */
-const parseTime = (time: number, locale: string, timezone: string): string => {
-  const dateTime = new Date(0);
-  dateTime.setUTCSeconds(time);
-
-  return dateTime
-    .toLocaleString(locale, { timeZone: timezone })
-    .split(",")
-    .pop() as string;
-};
-
-/**
- * Converts meters to miles.
- * @param {number} meters - The distance in meters.
- * @returns {number} The distance in miles.
- * @example
- * getMiles(1609.34); // 1
- * getMiles(8046.72); // 5
- * getMiles(32186.9); // 20
- */
-const getMiles = (meters: number): number => meters * MILES_CONVERSION_FACTOR;
-
-/**
  * Returns the day of the week for a given item.
  * @param {Object} item - The item object containing the dt property.
  * @returns {string} The day of the week.
@@ -161,58 +79,6 @@ const dayOfWeek = (item: { dt: number }): string =>
     .slice(0, SLICE_END_INDEX)
     .join(" ")
     .trim();
-
-/**
- * Converts millimeters to inches.
- * @param {number} data - The value in millimeters to be converted.
- * @returns {number} The converted value in inches.
- * @example
- * mmToInches(25.4); // 1
- * mmToInches(50.8); // 2
- * mmToInches(76.2); // 3
- * mmToInches(101.6); // 4
- */
-const mmToInches = (data: number): number => data / MM_TO_INCHES;
-
-/**
- * Generates a unique key for an alert.
- *
- * @param alert - The alert object.
- * @returns The generated key.
- */
-const generateAlertKey = (alert: Alert): string => {
-  return `${alert.sender_name}_${alert.start}_${alert.end}_${alert.tags.join(
-    "_"
-  )}`;
-};
-
-/**
- * Processes an array of alerts and returns an array of unique alerts.
- *
- * @param alerts - The array of alerts to process.
- * @returns An array of unique alerts.
- */
-const processAlerts = (alerts: Alert[]): Alert[] => {
-  const uniqueAlertMap = new Map<string, Alert>();
-
-  alerts.forEach((alert) => {
-    const key = generateAlertKey(alert);
-    if (!uniqueAlertMap.has(key)) {
-      uniqueAlertMap.set(key, alert);
-    }
-  });
-
-  return Array.from(uniqueAlertMap.values());
-};
-
-/**
- * Formats a timestamp into a string representation.
- * @param timestamp - The timestamp to format.
- * @returns The formatted string representation of the timestamp.
- */
-const formatDate = (timestamp: number) => {
-  return format(new Date(timestamp * MILLISECONDS_PER_SECOND), "p 'on' PPP");
-};
 
 /**
  * Formats the weather data for a given item.
@@ -300,31 +166,15 @@ const getVisibility = (visibility: number) => {
     : visibilityMiles;
 };
 
-/**
- * Returns the wind gust value, rounded down to the nearest integer.
- * If the wind gust value is undefined or less than 0, it returns 0.
- *
- * @param windGust - The wind gust value.
- * @returns The rounded down wind gust value.
- */
-const getWindGust = (windGust?: number) =>
-  typeof windGust === "undefined" || windGust < 0 ? 0 : Math.floor(windGust);
-
 export {
   dayOfWeek,
-  formatDate,
   formatWeatherData,
-  generateAlertKey,
   getMiles,
   getMoonPhase,
   getVisibility,
-  getWindDirection,
-  getWindGust,
   mmToInches,
   parseCity,
   parseTime,
   parseWeatherData,
-  processAlerts,
-  toCelsius,
-  toKph
+  toCelsius
 };
